Website fully functional

+++ api/src/main/java/org/savvas/milked/service/GroupInviteService.java
@@ -15,9 +15,6 @@ public class GroupInviteService {
     private MilkedUserRepository milkedUserRepository;
     private MilkingGroupRepository milkingGroupRepository;
     private UserService userService;
-    private static final String EMAIL_PATTERN =
-            "^[_A-Za-z0-9-\\+]+(\\.[_A-Za-z0-9-]+)*@"
-                    + "[A-Za-z0-9-]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$";
 
     @Autowired
     public GroupInviteService(GroupInviteRepository groupInviteRepository, MilkedUserRepository milkedUserRepository, MilkingGroupRepository milkingGroupRepository, UserService userService) {
@@ -60,9 +57,12 @@ public class GroupInviteService {
     public MilkingGroup acceptGroupInvite(String uuid, Long groupId) {
         MilkingGroup group = milkingGroupRepository.findOne(groupId);
         MilkedUser user = milkedUserRepository.findByUuid(uuid);
+        GroupInvite invite = groupInviteRepository.findOneByGroupIdAndUserId(groupId, user.getId());
+        if (invite == null) {
+            throw new ValidationException("Cannot find invite");
+        }
         group.addUserToGroup(user);
         milkingGroupRepository.save(group);
-        GroupInvite invite = groupInviteRepository.findOneByGroupIdAndUserId(groupId, user.getId());
         groupInviteRepository.delete(invite);
         return group;
     }
@@ -70,9 +70,12 @@ public class GroupInviteService {
     public MilkingGroup acceptGroupInviteForAnonymous(String uuid, Long groupId) {
         MilkingGroup group = milkingGroupRepository.findOne(groupId);
         MilkedUser user = milkedUserRepository.findByUuid(uuid);
+        GroupInvite invite = groupInviteRepository.findOneByGroupIdAndUserId(groupId, user.getId());
+        if (invite == null) {
+            throw new ValidationException("Cannot find invite");
+        }
         group.addUserToGroup(user);
         milkingGroupRepository.save(group);
-        GroupInvite invite = groupInviteRepository.findOneByGroupIdAndUserId(groupId, user.getId());
         groupInviteRepository.delete(invite);
         return group;
     }
